WIP: ExpressionTypeDescender
If we know [[a - b]] has integer type, and [[b]] is a pointer type, then [[a]] has to be a pointer type too.

This leads to some nice improvements in the code, but we also see a lot of new unions. It seems at first inspection that these are caused by low-level hacks "infecting" the inferred type system. Adding user declaration may alleviate this, by adding "firewalls" isolating hand-written assembly language hacks from the rest of a well-behaved program, but we want to have a discussion about how to best control this sort of thing, and whether we can help Reko detect this sort of usage and limit its spread.

Note: regression tests have not been run.